+++ packages/template-retail-react-app/app/pages/checkout-one-click/partials/one-click-checkout-header.test.js
@@ -7,7 +7,7 @@
 import React from 'react'
 import {screen} from '@testing-library/react'
 
-import CheckoutHeader from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/checkout-header'
+import CheckoutHeader from '@salesforce/retail-react-app/app/pages/checkout-one-click/partials/one-click-checkout-header'
 import {renderWithProviders} from '@salesforce/retail-react-app/app/utils/test-utils'
 
 test('renders component', () => {
