+++ packages/template-retail-react-app/app/pages/checkout-one-click/partials/one-click-payment-form.jsx
